refactor(server): extract the bind/drain/reload loop
Move run_with_config* and the SIGHUP / bootstrap accept loop
into server/listen.rs. Public gigastt::server entry points
are re-exported unchanged.
